Update LICENSE
License Description:

This application is patented and trademarked, and its use is strictly governed by the terms outlined in this license agreement. The intellectual property rights associated with this application, including all patents, trademarks, copyrights, and trade secrets, are owned exclusively by Techonward Ltd. As such, this application is protected under applicable national and international laws.

Restrictions on Use:

Non-Transferable License: The license granted for this application is non-transferable, meaning it cannot be shared, sold, sublicensed, or otherwise transferred to any third party.
Prohibited Use: No individual or entity is authorized to use, distribute, modify, reproduce, or reverse-engineer the application in any way. Unauthorized use of the application is strictly prohibited and may result in legal action.
Patent and Trademark Protections: The patented and trademarked aspects of the application cannot be reproduced or utilized by any other party without explicit, written permission from the intellectual property holder.
Exclusive Rights: The patent and trademark associated with this application grant exclusive rights to the owner, prohibiting any third-party development, modification, or use that could infringe upon the patented or trademarked features.
By using this application, you acknowledge and agree to comply with all terms and restrictions outlined in this agreement. Any violation of these terms will result in immediate termination of your license and may lead to legal consequences.